Fixed errors in the tests caused due to moving of the files to a new folder
